# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Two namespaces are documented: the original `/api/v1` surface and the
    consistent `/api/v2` surface introduced in issue #144. v2 is additive; v1 is
    unchanged. See docs/api-v2.md for the field-by-field v1 -> v2 mapping.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
  /api/v1/invocations/{tx_hash}/trace:
    get:
      tags:
        - Contracts
      summary: Cross-contract call graph trace for a transaction
      operationId: getInvocationTrace
      parameters:
        - name: tx_hash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trace returned
  /api/v1/watchdog/subscriptions/{id}/signing-secret:
    get:
      tags:
        - Watchdog
      summary: Reveal a subscription signing secret
      operationId: getSubscriptionSigningSecret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Secret returned
  /api/v1/watchdog/subscriptions/{id}/rotate:
    post:
      tags:
        - Watchdog
      summary: Rotate a subscription signing secret
      operationId: rotateSubscriptionSigningSecret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New secret returned
...
  /api/v1/stats/global:
    get:
      tags: [Stats]
      summary: Global index statistics
      operationId: getGlobalStats
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Global counts across all tracked contracts
          content:
            application/json:
              schema:
                type: object
                required: [tracked_contracts, total_events, total_invocations, total_storage_entries]
                properties:
                  tracked_contracts:
                    type: integer
                    format: int64
                  total_events:
                    type: integer
                    format: int64
                  total_invocations:
                    type: integer
                    format: int64
                  total_storage_entries:
                    type: integer
                    format: int64
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts:
    get:
      tags: [Contracts]
      summary: List tracked contracts
      operationId: listContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by contract status (pending, active, backfilling, error, paused).
        - $ref: "#/components/parameters/TagParam"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [id, label, network, status, added_at]
          description: Column to sort by. Defaults to id when omitted; values outside the enum are rejected with 422.
        - name: dir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction. Defaults to asc when omitted; values outside the enum are rejected with 422.
      responses:
        "200":
          description: Page of contracts
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contract"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Register a contract for tracking
      operationId: registerContract
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, network]
              properties:
                id:
                  type: string
                  pattern: "^C[A-Za-z0-9]{55}$"
                  description: Stellar contract ID (56 chars, starts with C).
                network:
                  type: string
                  enum: [testnet, mainnet, futurenet, standalone]
                label:
                  type: string
                  description: Optional human-friendly label.
      responses:
        "201":
          description: Contract registered (status pending until backfill completes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/batch:
    post:
      tags: [Contracts]
      summary: Apply a bulk action to many contracts
      operationId: batchContracts
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchContractsRequest"
      responses:
        "200":
          description: Bulk action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchContractsResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/tags:
    post:
      tags: [Contracts]
      summary: Add a tag to a contract
      description: |
        Adds a user-defined tag to a contract. The tag is trimmed and
        lowercased before storage. Adding a tag that is already present is a
        no-op, so the operation is idempotent.
      operationId: addContractTag
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractTagRequest"
      responses:
        "200":
          description: The contract's full, sorted tag list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractTags"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/tags/{tag}:
    delete:
      tags: [Contracts]
      summary: Remove a tag from a contract
      description: |
        Removes a user-defined tag from a contract. Removing a tag that is not
        present is a no-op.
      operationId: removeContractTag
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/TagPath"
      responses:
        "204":
          description: Tag removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/notes:
    get:
      tags: [Contracts]
      summary: List a contract's notes
      description: |
        Returns a contract's markdown notes, newest first. Notes are open to
        read like the rest of the v0.1 surface; a contract with no notes gets
        an empty list rather than an error.
      operationId: listContractNotes
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract's notes, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNoteList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Add a note to a contract
      description: |
        Appends a markdown note to a contract. The stored author is the caller
        identity (X-User-ID), which is what later authorizes editing and
        deleting the note. Authoring requires the contributor role.
      operationId: createContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractNoteRequest"
      responses:
        "201":
          description: The stored note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/notes/{noteID}:
    patch:
      tags: [Contracts]
      summary: Rewrite a note
      description: |
        Replaces a note's body. Only the note's author may rewrite it; any
        other contributor gets 403. The stored author is never reassigned.
      operationId: updateContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NoteID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractNoteRequest"
      responses:
        "200":
          description: The updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Contracts]
      summary: Delete a note
      description: |
        Removes a note. Only the note's author may delete it. Deleting a note
        that is already gone is a no-op, so a retried delete succeeds.
      operationId: deleteContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NoteID"
      responses:
        "204":
          description: Note deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v2/stats/global:
    get:
      tags: [API v2]
      summary: Network-wide summary
      operationId: v2GlobalStats
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Global statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/V2GlobalStats"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist and contract-group endpoints read the user ID from X-User-ID.
...